+++ great_international/models/great_international.py
@@ -1724,6 +1724,11 @@ class AboutUkRegionPage(panels.AboutUkRegionPagePanels, BaseInternationalPage):
         related_name='+',
         blank=True
     )
+    region_summary_section_strapline = models.TextField(
+        max_length=255,
+        blank=True,
+        help_text="Displayd above Region Section Summary Intro"
+    )
     region_summary_section_intro = models.TextField(max_length=255, blank=True)
     region_summary_section_content = MarkdownField(blank=True)